Updated the AutomationTriggerType
